LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.meta/worker-src-none/sharedworker-classic.https-expected.txt


PASS Content Security Policy: Expects blocked for sharedworker-classic to same-https origin and keep-origin redirection from https context.
FAIL Content Security Policy: Expects blocked for sharedworker-classic to same-https origin and keep-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for sharedworker-classic to same-https origin and no-redirect redirection from https context.
FAIL Content Security Policy: Expects blocked for sharedworker-classic to same-https origin and no-redirect redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
